+++ Changelog.md
@@ -21,6 +21,7 @@
 - Added the cart item percent discount promotion action type
 - Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items
 - Added the `priority` and `subtitle` fields to the `products`, `master_products` and `master_product_variants` tables
+- Added the `getListableVariants()` method / `$listable_variants` property to the `MasterProduct` model
 - Added default sorting to `ProductSearch`: by priority, descending
 - Added the static `getAvailableWithoutCoupon()` and `getAvailableOnes()` methods to the `Promotion` model (but not to the interface)
 - Added the `active()` and `notDepeleted()` scopes to the `Promotion` model