+++ src/MasterProduct/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2025-XX-YY
 
 - Added the `priority` and `subtitle` fields to the `master_products` and `master_product_variants` tables
+- Added the `getListableVariants()` method / `$listable_variants` property to the `MasterProduct` model
 
 ## 5.0.0
 ##### 2025-09-03